Notification types
Sveltos and k8s-cleaner shares some notification types (Slack, Webex,
Discord, Teams and soon SMTP). But other notifications types are different
(Reports and soon Telegram).

I feel the right approach is to:

1. have common logic in libsveltos but
2. have different Notification Types

This will allow logic which is common to be be moved to libsveltos while
yet allowing different notification types to be supported.